Added support for hardware and firmware version requests. Added support to request the length of the serial number. The EEPROM slot function also has the ability to read serial number, but not the length. Trying to read the length throws an exception.

+++ src/api/seabreezeapi/OpticalBenchFeatureAdapter.cpp
@@ -64,8 +64,8 @@ OpticalBenchFeatureAdapter::~OpticalBenchFeatureAdapter() {
 #pragma warning(disable : 4101)// unreferenced local variable
 #endif
 
-unsigned int OpticalBenchFeatureAdapter::readOpticalBenchFiberDiameterMicrons(int *errorCode) {
-	unsigned int returnValue;
+unsigned short int OpticalBenchFeatureAdapter::readOpticalBenchFiberDiameterMicrons(int *errorCode) {
+	unsigned short int returnValue;
 
 	// no memory allocated, just pass it through
 	try {
@@ -78,8 +78,8 @@ unsigned int OpticalBenchFeatureAdapter::readOpticalBenchFiberDiameterMicrons(in
 	return returnValue;
 }
 
-unsigned int OpticalBenchFeatureAdapter::readOpticalBenchSlitWidthMicrons(int *errorCode) {
-	unsigned int returnValue;
+unsigned short int OpticalBenchFeatureAdapter::readOpticalBenchSlitWidthMicrons(int *errorCode) {
+	unsigned short int returnValue;
 
 	// no memory allocated, just pass it through
 	try {
